WEA-27 Move entity forecast model to seperate package
