+++ app/src/main/java/com/ayush/weatherapp/home/TabPagerAdapter.java
@@ -4,8 +4,8 @@ import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentStatePagerAdapter;
-import com.ayush.weatherapp.entities.DailyDataEntity;
-import com.ayush.weatherapp.entities.HourlyDataEntity;
+import com.ayush.weatherapp.entities.forecast.DailyDataEntity;
+import com.ayush.weatherapp.entities.forecast.HourlyDataEntity;
 import java.util.List;
 
 public class TabPagerAdapter extends FragmentStatePagerAdapter {
