+++ app/src/main/java/com/ayush/weatherapp/mapper/ForecastRealmToEntityMapper.java
@@ -1,11 +1,11 @@
 package com.ayush.weatherapp.mapper;
 
-import com.ayush.weatherapp.entities.CurrentForecastEntity;
-import com.ayush.weatherapp.entities.DailyDataEntity;
-import com.ayush.weatherapp.entities.DailyForecastEntity;
-import com.ayush.weatherapp.entities.ForecastEntity;
-import com.ayush.weatherapp.entities.HourlyDataEntity;
-import com.ayush.weatherapp.entities.HourlyForecastEntity;
+import com.ayush.weatherapp.entities.forecast.CurrentForecastEntity;
+import com.ayush.weatherapp.entities.forecast.DailyDataEntity;
+import com.ayush.weatherapp.entities.forecast.DailyForecastEntity;
+import com.ayush.weatherapp.entities.forecast.ForecastEntity;
+import com.ayush.weatherapp.entities.forecast.HourlyDataEntity;
+import com.ayush.weatherapp.entities.forecast.HourlyForecastEntity;
 import com.ayush.weatherapp.realm.model.forecast.CurrentForecast;
 import com.ayush.weatherapp.realm.model.forecast.DailyData;
 import com.ayush.weatherapp.realm.model.forecast.DailyForecast;
