+++ app/src/main/java/com/ayush/weatherapp/repository/weather/WeatherRepository.java
@@ -1,6 +1,6 @@
 package com.ayush.weatherapp.repository.weather;
 
-import com.ayush.weatherapp.entities.ForecastEntity;
+import com.ayush.weatherapp.entities.forecast.ForecastEntity;
 import io.reactivex.Observable;
 
 public interface WeatherRepository {
